# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Bookly",
        "description": "\nA REST API for a book review web service.\n\nThis REST API is able to;\n- Create Read Update And delete books\n- Add reviews to books\n- Add tags to Books e.t.c.\n    ",
        "termsOfService": "httpS://example.com/tos",
        "contact": {
            "name": "Ssali Jonathan",
            "url": "https://github.com/jod35",
            "email": "ssalijonathank@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/license/mit"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/items/": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get All Items",
                "operationId": "readItems",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1000,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemsPublicModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "items"
                ],
                "summary": "Create Item",
                "operationId": "createItem",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemCreateModel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{item_uid}": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get Item",
                "operationId": "readItem",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Uid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "items"
                ],
                "summary": "Update Item",
                "operationId": "updateItem",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Uid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemUpdateModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "items"
                ],
                "summary": "Delete Item",
                "operationId": "deleteItem",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Uid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{item_uid}/ingredients": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get Item Ingredients",
                "operationId": "readItemIngredients",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Uid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientsTreeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/{item_uid}/possibilities": {
            "get": {
                "tags": [
                    "items"
                ],
                "summary": "Get Item Possibilities",
                "operationId": "readItemPossibilities",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "item_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Uid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PossibilitiesResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/items/links": {
            "post": {
                "tags": [
                    "items"
                ],
                "summary": "Create Item Link",
                "operationId": "createItemLink",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ItemLinkCreateModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemLinkModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ]
            }
        },
        "/api/v1/items/links/{link_uid}": {
            "delete": {
                "tags": [
                    "items"
                ],
                "summary": "Delete Item Link",
                "operationId": "deleteItemLink",
                "security": [
                    {
                        "AccessTokenBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "link_uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Link Uid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IngredientsTreeResponse": {
                "properties": {
                    "root": {
                        "$ref": "#/components/schemas/RecipeTreeNode"
                    }
                },
                "type": "object",
                "required": [
                    "root"
                ],
                "title": "IngredientsTreeResponse"
            },
...
            "ItemCreateModel": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "rarity": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rarity"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "max_drop": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Drop"
                    },
                    "type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Type"
                    },
                    "chi": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chi"
                    },
                    "texture_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Texture Type"
                    },
                    "collision_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Collision Type"
                    },
                    "seed_color": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seed Color"
                    },
                    "grow_time": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grow Time"
                    },
                    "default_gems_drop": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Default Gems Drop"
                    },
                    "hits_with_hand": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hits With Hand"
                    },
                    "hits_with_pickaxe": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hits With Pickaxe"
                    },
                    "restore_time_seconds": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Restore Time Seconds"
                    },
                    "scraped": {
                        "type": "boolean",
                        "title": "Scraped",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "ItemCreateModel"
            },
...
            "ItemModel": {
                "properties": {
                    "uid": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Uid"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "rarity": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rarity"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "max_drop": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Max Drop"
                    },
                    "type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Type"
                    },
                    "chi": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chi"
                    },
                    "texture_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Texture Type"
                    },
                    "collision_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Collision Type"
                    },
                    "seed_color": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Seed Color"
                    },
                    "grow_time": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grow Time"
                    },
                    "default_gems_drop": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Default Gems Drop"
                    },
                    "hits_with_hand": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hits With Hand"
                    },
                    "hits_with_pickaxe": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Hits With Pickaxe"
                    },
                    "restore_time_seconds": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Restore Time Seconds"
                    },
                    "scraped": {
                        "type": "boolean",
                        "title": "Scraped",
                        "default": false
                    },
                    "created_by_uid": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Created By Uid"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "update_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Update At"
                    }
                },
                "type": "object",
                "required": [
                    "uid",
                    "name",
                    "created_at",
                    "update_at"
                ],
                "title": "ItemModel"
            },
...
            "ItemsPublicModel": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/ItemModel"
                        },
                        "type": "array",
                        "title": "Data"
                    },
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    }
                },
                "type": "object",
                "required": [
                    "data",
                    "count"
                ],
                "title": "ItemsPublicModel"
            },
...
            "PossibilitiesResponse": {
                "properties": {
                    "possibilities": {
                        "items": {
                            "$ref": "#/components/schemas/ItemModel"
                        },
                        "type": "array",
                        "title": "Possibilities"
                    }
                },
                "type": "object",
                "required": [
                    "possibilities"
                ],
                "title": "PossibilitiesResponse"
            },
...
        },
        "securitySchemes": {
            "AccessTokenBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "RefreshTokenBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}